Cleanup Grin main binary crate
Split up each subcommand into its own file under the `cmd`
subcrate.
